Adds MIDI Thru suppourt
Incoming MIDI CC will now be passed through as if the knob was being turned.

Previously sending a MIDI CC to the MFT for a knob would update the LED, but the MFT would not pass that value through. This now passes all incoming MIDI through enabling the device to be updated, LED changes, and the devices sends that MIDI value to it's MIDI out.